requirements.txt
fastapi
uvicorn[standard]
python-multipart
jinja2
mysql-connector-python
pydantic
itsdangerous>=2.1.2
python-dotenv